requirements.txt

lxml
httpx
python-dotenv
ddddocr
